fix (res bonus buildings) fixed when bot tried to add res bonus building to the task list, it removed the task running so there were multiple tasks executing simultaneously (caused errors)

+++ TbsCore/Parsers/InfrastructureParser.cs
@@ -65,7 +65,7 @@ namespace TravBotSharp.Files.Parsers
                 ret.Add(new BuildingCurrently()
                 {
                     Duration = DateTime.Now.Add(duration),
-                    Level = (int)Parser.RemoveNonNumeric(level),
+                    Level = (byte)Parser.RemoveNonNumeric(level),
                 });
             }
             return ret;
